feat: add full RAG functionality
Add program to use the vector db to conduct a RAG.
First, conduct a similarity search between a query
and the pdf chunks. Then, conduct a re-ranking.
Then, use these context chunks as input to an LLM
to answer the question. Finally, it is possible to
review the sections cited by the LLM to ensure that
the model has indeed answered the question correctly.